move SBO cache to separate file, remove AnyCache
